# java/swagger/MC-Observability Swagger(OpenAPI 3.0).yaml
openapi: 3.0.1
info:
  title: MC-Observability API Documentation
  description: API documentation for MC-Observability
  version: 0.4.1
servers:
- url: localhost:18080
...
paths:
...
  /api/o11y/trigger/history:
    get:
      tags:
      - "[Trigger] Monitoring Measurement Trigger"
      summary: GetPaginatedTriggerHistories
      description: Get paginated trigger histories
      operationId: GetPaginatedTriggerHistories
      parameters:
      - name: page
        in: query
        description: page number (1 .. N)
        required: false
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        description: size of page (1 .. N)
        required: false
        schema:
          type: integer
          format: int32
      - name: sortBy
        in: query
        description: sort by properties
        required: false
        schema:
          type: string
      - name: sortDirection
        in: query
        description: "sort direction (asc, desc)"
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggerHistoryPageResponse'
              examples:
                GetPaginatedTriggerHistories:
                  value: |-
                    {
                      "content" : [ ],
                      "pageable" : {
                        "pageNumber" : 0,
                        "pageSize" : 1,
                        "sort" : {
                          "sorted" : false,
                          "empty" : true,
                          "unsorted" : true
                        },
                        "offset" : 0,
                        "paged" : true,
                        "unpaged" : false
                      },
                      "totalPages" : 1,
                      "totalElements" : 0,
                      "numberOfElements" : 0
                    }
...
  /api/o11y/trigger/alert/test-history:
    get:
      tags:
      - "[Trigger - Only Developer] Monitoring Trigger Event Handler"
      summary: GetPaginatedAlertTestHistories
      description: Get paginated alert test histories
      operationId: GetPaginatedAlertTestHistories
      parameters:
      - name: page
        in: query
        description: page number (1 .. N)
        required: false
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        description: size of page (1 .. N)
        required: false
        schema:
          type: integer
          format: int32
      - name: sortBy
        in: query
        description: sort by properties (id..)
        required: false
        schema:
          type: string
      - name: direction
        in: query
        description: "sort direction (asc, desc)"
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertTestHistoryPageResponse'
              examples:
                GetPaginatedAlertTestHistories:
                  value: |-
                    {
                      "content" : [ ],
                      "pageable" : {
                        "pageNumber" : 0,
                        "pageSize" : 1,
                        "sort" : {
                          "sorted" : false,
                          "empty" : true,
                          "unsorted" : true
                        },
                        "offset" : 0,
                        "paged" : true,
                        "unpaged" : false
                      },
                      "totalPages" : 1,
                      "totalElements" : 0,
                      "numberOfElements" : 0
                    }
...
  /api/o11y/monitoring/{nsId}/{mciId}/vm:
    get:
      tags:
      - "[Manager] Monitoring VM Management"
      summary: GetVMByNsMci
      description: Retrieve target list by NS/MCI
      operationId: GetVMByNsMci
      parameters:
      - name: nsId
        in: path
        description: "nsId (e.g., ns-1)"
        required: true
        schema:
          type: string
      - name: mciId
        in: path
        description: "mciId (e.g., mci-1)"
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VMDTO'
              examples:
                GetVMByNsMci:
                  value: |-
                    {
                      "rs_code" : "0000",
                      "rs_msg" : "success",
                      "data" : [ {
                        "name" : "mcmp-vm",
                        "description" : "string",
                        "vm_id" : "vm-1",
                        "influx_seq" : 0,
                        "vm_status" : "RUNNING",
                        "monitoring_service_status" : "ACTIVE",
                        "log_service_status" : "ACTIVE",
                        "ns_id" : "ns-1",
                        "mci_id" : "mci-1"
                      } ],
                      "error_message" : ""
                    }
...
  /api/o11y/trigger/alert/alerts/search:
    get:
      tags:
      - "[Trigger - Only Developer] Monitoring Trigger Event Handler"
      summary: SearchAlertsByTitle
      description: Search alerts by title
      operationId: SearchAlertsByTitle
      parameters:
      - name: title
        in: query
        description: alert title to search
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'
              examples:
                SearchAlertsByTitle:
                  value: "[ ]"
  /api/o11y/trigger/history/{id}/comment:
    put:
      tags:
      - "[Trigger] Monitoring Measurement Trigger"
      summary: UpdateTriggerHistoryComment
      description: Update trigger history comment
      operationId: UpdateTriggerHistoryComment
      parameters:
      - name: id
        in: path
        description: trigger history id
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/TriggerHistoryCommentUpdateRequest'
            examples:
              UpdateTriggerHistoryComment:
                value: |-
                  {
                    "comment" : "Alert resolved - false positive"
                  }
      responses:
        "200":
          description: "200"
...
  /api/o11y/monitoring/{nsId}/{mciId}/vm/{vmId}:
    get:
      tags:
      - "[Manager] Monitoring VM Management"
      summary: GetVM
      description: Retrieve single target
      operationId: GetVM
      parameters:
      - name: nsId
        in: path
        description: "nsId (e.g., ns-1)"
        required: true
        schema:
          type: string
      - name: mciId
        in: path
        description: "mciId (e.g., mci-1)"
        required: true
        schema:
          type: string
      - name: vmId
        in: path
        description: "vmId (e.g., vm-1)"
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VMDTO'
              examples:
                GetVM:
                  value: |-
                    {
                      "rs_code" : "0000",
                      "rs_msg" : "success",
                      "data" : {
                        "name" : "mcmp-vm",
                        "description" : "string",
                        "vm_id" : "vm-1",
                        "influx_seq" : 0,
                        "vm_status" : "RUNNING",
                        "monitoring_service_status" : "ACTIVE",
                        "log_service_status" : "ACTIVE",
                        "ns_id" : "ns-1",
                        "mci_id" : "mci-1"
                      },
                      "error_message" : ""
                    }
    put:
      tags:
      - "[Manager] Monitoring VM Management"
      summary: PutVM
      description: Update target
      operationId: PutVM
      parameters:
      - name: nsId
        in: path
        description: "nsId (e.g., ns-1)"
        required: true
        schema:
          type: string
      - name: mciId
        in: path
        description: "mciId (e.g., mci-1)"
        required: true
        schema:
          type: string
      - name: vmId
        in: path
        description: "vmId (e.g., vm-1)"
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/VMRequestDTO'
            examples:
              PutVM:
                value: |-
                  {
                    "name" : "mcmp-vm",
                    "description" : "string"
                  }
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VMDTO'
              examples:
                PutVM:
                  value: |-
                    {
                      "rs_code" : "0000",
                      "rs_msg" : "success",
                      "data" : {
                        "name" : "mcmp-vm",
                        "description" : "string",
                        "vm_id" : "vm-1",
                        "influx_seq" : 0,
                        "vm_status" : "RUNNING",
                        "monitoring_service_status" : "ACTIVE",
                        "log_service_status" : "ACTIVE",
                        "ns_id" : "ns-1",
                        "mci_id" : "mci-1"
                      },
                      "error_message" : ""
                    }
    post:
      tags:
        - "[Manager] Monitoring VM Management"
      summary: PostVM
      description: Create target (VM)
      operationId: PostVM
      parameters:
        - name: nsId
          in: path
          description: "Namespace ID (e.g., ns-1)"
          required: true
          schema:
            type: string
        - name: mciId
          in: path
          description: "MCI ID (e.g., mci-1)"
          required: true
          schema:
            type: string
        - name: vmId
          in: path
          description: "VM ID (e.g., vm-1)"
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/VMRequestDTO'
            examples:
              PostVM:
                value: |-
                  {
                    "name" : "mcmp-vm",
                    "description" : "string"
                  }
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResBody'
              examples:
                PostVM:
                  value: |-
                    {
                      "rs_code" : "0000",
                      "rs_msg" : "success",
                      "data" : null,
                      "error_message" : ""
                    }
    delete:
      tags:
      - "[Manager] Monitoring VM Management"
      summary: DeleteVM
      description: Delete target
      operationId: DeleteVM
      parameters:
      - name: nsId
        in: path
        description: "nsId (e.g., ns-1)"
        required: true
        schema:
          type: string
      - name: mciId
        in: path
        description: "mciId (e.g., mci-1)"
        required: true
        schema:
          type: string
      - name: vmId
        in: path
        description: "vmId (e.g., vm-1)"
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api-o11y-monitoring-nsId-mciId-vm-vmId-357262848'
              examples:
                DeleteVM:
                  value: |-
                    {
                      "rs_code" : "0000",
                      "rs_msg" : "success",
                      "data" : null,
                      "error_message" : ""
                    }
...
components:
  schemas:
...
    VMRequestDTO:
      title: VMRequestDTO
      type: object
      properties:
        name:
          type: string
          description: "Target name(e.g.,  mcmp-vm)"
        description:
          type: string
          description: Description
...
    TriggerHistoryPageResponse:
      title: TriggerHistoryPageResponse
      type: object
      properties:
        numberOfElements:
          type: number
          description: number of elements
        totalPages:
          type: number
          description: total pages
        pageable:
          type: object
          description: specific page info
        content:
          type: array
          description: trigger history list
          items:
            oneOf:
            - type: object
            - type: boolean
            - type: string
            - type: number
        totalElements:
          type: number
          description: total elements
...
    VMDTO:
      title: VMDTO
      type: object
      properties:
        error_message:
          type: string
          description: Error message
        data:
          type: array
          description: Target information list
          items:
            type: object
            properties:
              influx_seq:
                type: number
                description: Influx sequence
              vm_id:
                type: string
                description: vmId
              mci_id:
                type: string
                description: "mciId (e.g.,  ci-1)"
              vm_status:
                type: string
                description: Target status
                enum:
                - RUNNING
                - FAILED
              log_service_status:
                type: string
                description: Log agent service status
                enum:
                - ACTIVE
                - INACTIVE
                - ACTIVATING
                - DEACTIVATING
                - RELOADING
                - NOT_EXIST
                - FAILED
              ns_id:
                type: string
                description: "nsId (e.g., ns-1)"
              monitoring_service_status:
                type: string
                description: Monitoring agent service status
                enum:
                - ACTIVE
                - INACTIVE
                - ACTIVATING
                - DEACTIVATING
                - RELOADING
                - NOT_EXIST
                - FAILED
              name:
                type: string
                description: "Target name (e.g., mcmp-vm)"
              description:
                type: string
                description: Description
        rs_code:
          type: string
          description: "Response code (e.g., 0000)"
        rs_msg:
          type: string
          description: "Response message (e.g., Success)"
...
    AlertResponse:
      type: array
      description: List of matching alerts
      items:
        oneOf:
        - type: object
        - type: boolean
        - type: string
        - type: number
...
    TriggerHistoryCommentUpdateRequest:
      title: TriggerHistoryCommentUpdateRequest
      type: object
      properties:
        comment:
          type: string
          description: comment to update
...
    ResBody:
      type: object
      properties:
        error_message:
          type: string
          description: Error message
        data:
          type: object
          description: Response data (nullable)
        rs_code:
          type: string
          description: "Response code (e.g., 0000)"
        rs_msg:
          type: string
          description: "Response message (e.g., Success)"
...
    api-o11y-monitoring-nsId-mciId-vm-vmId-357262848:
      type: object
      properties:
        error_message:
          type: string
          description: Error message
        rs_code:
          type: string
          description: "Response code (e.g., 0000)"
        rs_msg:
          type: string
          description: "Response message (e.g., Success)"
...
    AlertTestHistoryPageResponse:
      title: AlertTestHistoryPageResponse
      type: object
      properties:
        numberOfElements:
          type: number
          description: number of elements
        totalPages:
          type: number
          description: total pages
        pageable:
          type: object
          description: specific page info
        content:
          type: array
          description: alert test history list
          items:
            oneOf:
            - type: object
            - type: boolean
            - type: string
            - type: number
        totalElements:
          type: number
          description: total elements
...